3-examples/01_customer_service/pyproject.toml
[project]
name = "01-customer-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "logfire>=3.15.0",
    "openai-agents>=0.0.14",
    "python-dotenv>=1.1.0",
]
